requirements.txt
fastapi
strawberry-graphql
uvicorn
pymongo
python-dotenv
